pyproject.toml
[tool.poetry]
name = "btc-perp-trader"
version = "1.0.0"
description = "Robô de trading algorítmico para BTC-PERP"
authors = ["Manus <manus@example.com>"]
packages = [
  { include = "btc_perp_trader", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
ccxt = "^4.0.0"
cryptofeed = "^0.22.0"
aiohttp = "^3.9.5"
pandas = "^2.0.0"
polars = "^0.19.0"
duckdb = "^0.9.0"
pyarrow = "^14.0.0"
tqdm      = "^4.66.4"
ta = "^0.10.0"
numba = ">=0.56,<0.57"
scikit-learn = "^1.4"
xgboost = "^2.0"
lightgbm = "^4.0.0"
river      = "0.22.0"
transformers = "^4.41.0"
newspaper3k = "^0.2.8"
torch = { version = "^2.2", optional = true }
optuna = "^3.6"
joblib = "^1.4"
torchmetrics = "^0.11.0"
darts = "^0.20.0"
vectorbt = "^0.25.4"
plotly = "^5.22"
flask = "^2.3.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.0"
backtrader = "^1.9.0"
empyrical = "^0.5.0"
riskfolio-lib = "^4.0.0"
FastAPI = "^0.104.0"
pydantic = "^2.0.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
python-dotenv = "^1.0"
websockets = "^12.0"
streamlit = "^1.35"
nltk = "^3.9.1"
vaderSentiment = "^3.3.2"
requests = "^2.32.3"   # para CryptoPanic / Fear&Greed
orjson = "^3.10.0"        # grava/ler cache de manchetes muito rápido
pandas-ta = {version = "0.3.14b0", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.0"
black = "^23.0.0"
mypy = "^1.0.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
generate-report = "btc_perp_trader.backtest.generate_report:main"
ping-binance = "scripts.ping_test:main"
btc-trader = "btc_perp_trader.cli:cli"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true
exclude = ["robo_btc_api"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


